build(SUP-3465): use generic packages in admin servers
Adds `default_platform` variable to project N config in admin servers,
so that a single package without the setting can be used for either
AWS or GCP deployments.

Also removes setup script from repo, as the scripts in this repo are
out of date, and the actual source-of-truth script lives in the docs
site code.